+++ src/auth/password.rs
@@ -1,24 +1,44 @@
-// use argon2::{
-//     password_hash::{rand_core::OsRng, PasswordHash, PasswordHasher, PasswordVerifier, SaltString},
-//     Argon2,
-// };
-// use log::error;
-//
-// pub async fn hash(password: &[u8]) -> Option<String> {
-//     let salt = SaltString::generate(&mut OsRng);
-//     match Argon2::default().hash_password(password, &salt) {
-//         Ok(pwd) => Some(pwd.to_string()),
-//         Err(err) => {
-//             error!("Failed to hash password {}", err);
-//             None
-//         }
-//     }
-// }
-//
-// pub fn verify_password(hash: &str, password: &[u8]) -> bool {
-//     let parsed_hash = PasswordHash::new(hash)?;
-//     match Argon2::default().verify_password(password, &parsed_hash) {
-//         Ok(_) => true,
-//         Err(_) => false
-//     }
-// }
+use std::io::ErrorKind;
+use rand::{Rng, distributions::Alphanumeric, rngs::OsRng};
+
+fn generate_salt(length: usize) -> String {
+    let rng = OsRng;
+    let salt: String = rng
+        .sample_iter(&Alphanumeric)
+        .take(length)
+        .map(char::from)
+        .collect();
+    salt
+}
+
+pub(crate) fn hash(password: &[u8]) -> Option<String> {
+    let salt = generate_salt(64);
+    if password.len() > 0 {
+        let config = argon2::Config::default();
+        if let Ok(hash) = argon2::hash_encoded(password, salt.as_bytes(), &config) {
+            return Some(hash);
+        }
+    }
+    None
+}
+
+pub(crate) fn verify_password(hash: &str, password: &[u8]) -> bool {
+    if let Ok(valid) = argon2::verify_encoded(hash, password) {
+        return valid;
+    }
+    false
+}
+
+
+pub(crate) fn generate_password() -> std::io::Result<String> {
+    match rpassword::prompt_password("password> ") {
+        Ok(pwd) => {
+          match hash(pwd.as_bytes()) {
+              None => Err(std::io::Error::new(ErrorKind::Other, "Failed to generate hash")),
+              Some(hash) => Ok(hash),
+          }
+        },
+        Err(err) => Err(err)
+    }
+}
+