+++ CHANGELOG.md
@@ -14,7 +14,12 @@
 * Mapper can now set `epg_channel_id`.
 * Added environment variables for User Credentials `username`, `password` and `token` in format `${env:<EnvVarName>}` where `<EnvVarName>` should be replaced.
 * Added `web_ui_enabled` to `config.yml`. Default is `true`. Set to `false` to disable webui.
-* Added `web_auth_enabled` to `config.yml`. Default is `false`. Set to `true` to enable webui authentication.
+* Added `web_auth` to `config.yml` struct for web-ui-authentication is optional.
+   - `enabled`: default true
+   - `issuer` issuer for jwt token
+   - `secret` secret for jwt token
+   - `userfile` userfile with generated userfile in format "username: password" per file
+* Password generation argument --genpwd  to generate passwords for userfile. 
 
 # v1.1.8(2024-03-06)
 * Fixed WebUI Option-Select